refactor media details components to improve structure and image loading
