+++ app/src/main/java/com/berlin/aflami/di/usecaseModule.kt
@@ -7,22 +7,13 @@ import usecase.GetRecentHistoryUseCase
 import usecase.mediadetails.GetMovieDetailsUseCase
 import usecase.mediadetails.GetMovieGalleryUseCase
 import usecase.SearchByActorNameUseCase
-import usecase.GetMovieCastUseCase
 import usecase.GetMovieGenresUseCase
-import usecase.GetMovieReviewUseCase
 import usecase.mediadetails.GetMovieCastUseCase
 import usecase.mediadetails.GetMovieReviewUseCase
 import usecase.GetSearchMoviesUseCase
 import usecase.GetSearchTvShowsUseCase
 import usecase.SaveRecentHistoryUseCase
-import usecase.GetSeasonEpisodesUseCase
-import usecase.GetSimilarMoviesUseCase
-import usecase.GetSimilarSeriesUseCase
-import usecase.GetTvShowDetailsUseCase
-import usecase.GetSeriesGalleryUseCase
-import usecase.GetSeriesCastUseCase
 import usecase.GetSeriesGenresUseCase
-import usecase.GetSeriesReviewUseCase
 import usecase.mediadetails.GetSeasonEpisodesUseCase
 import usecase.mediadetails.GetSimilarMoviesUseCase
 import usecase.mediadetails.GetSimilarSeriesUseCase
