+++ presentation/ui/src/main/kotlin/com/berlin/aflami/screens/mediadetails/components/MediaCastItem.kt
@@ -1,31 +1,25 @@
 package com.berlin.aflami.screens.mediadetails.components
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import coil3.compose.AsyncImage
-import coil3.compose.AsyncImagePainter
-import coil3.compose.rememberAsyncImagePainter
 import com.berlin.aflami.ui.theme.Theme
-import com.berlin.ui.R
 
 @Composable
 fun MediaCastItem(
@@ -33,37 +27,26 @@ fun MediaCastItem(
     name: String,
     poster: String,
 ) {
-    val painter = rememberAsyncImagePainter(poster)
-    val state by painter.state.collectAsState()
     Column(
         verticalArrangement = Arrangement.spacedBy(2.dp)
     ) {
-
         Box(
             modifier = modifier
+                .size(48.dp)
                 .clip(RoundedCornerShape(16.dp))
-                .border(1.dp, color = Theme.color.stroke, RoundedCornerShape(16.dp)),
+                .border(1.dp, color = Theme.color.stroke,
+                    RoundedCornerShape(16.dp)),
+            contentAlignment = Alignment.Center
         ){
-            when (state) {
-                is AsyncImagePainter.State.Success ->
-                    AsyncImage(
-                        modifier = modifier,
-                        model = poster,
-                        contentDescription = stringResource(R.string.cast_image),
-                        contentScale = ContentScale.Crop
-                    )
-
-                else -> {
-                    Image(
-                        painter = painterResource(com.berlin.designsystem.R.drawable.ic_placeholder),
-                        contentDescription = null,
-                        modifier = Modifier
-                            .size(48.dp)
-                            .align(Alignment.Center),
-                        contentScale = ContentScale.Inside,
-                    )
-                }
-            }
+            AsyncImage(
+                model = poster,
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier.fillMaxSize(),
+                placeholder = painterResource(com.berlin.ui.R.drawable.place_holder),
+                error = painterResource(com.berlin.ui.R.drawable.place_holder),
+                fallback = painterResource(com.berlin.ui.R.drawable.place_holder),
+            )
         }
 
         Text(
@@ -73,7 +56,6 @@ fun MediaCastItem(
             maxLines = 1,
             overflow = TextOverflow.Ellipsis,
             modifier = Modifier.width(74.dp)
-
         )
     }
 }
